networked countdown and fixed match duration

+++ Assets/Scripts/GameplayClockUI.cs
@@ -10,7 +10,7 @@ public class GameplayClockUI : MonoBehaviour {
     private TimeSpan timeSpan;
 
     private void Update() {
-        timeSpan = TimeSpan.FromSeconds(DemoManager.Instance.GetGamePlayingTimer());
+        timeSpan = TimeSpan.FromSeconds((double)(DemoManager.Instance.GetGamePlayingTimer()));
         gameplayTimer.text = string.Format("{0:00}:{1:00}", timeSpan.Minutes, timeSpan.Seconds);
 
         if (Input.GetKeyDown(KeyCode.T)) {
